requirements.txt
asgiref==3.11.1
boto3==1.42.41
botocore==1.42.41
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==4.2.28
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
environ==1.0
idna==3.11
jmespath==1.1.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.11.0
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
